fix(train): drop a redundant cast in the ln_1p test
CI clippy caught `i as i32` where `i` was already inferred as i32.

Local clippy never reached tang-train to find this: tang-la fails first under
-D warnings on aarch64, where the NEON cfg block makes the scalar fallback
unreachable. That failure is pre-existing and inactive on CI's x86_64 runners,
but it does mean the workspace clippy command cannot be run end to end on an
Apple Silicon machine.

+++ crates/tang-train/src/ctc.rs
@@ -576,7 +576,7 @@ mod tests {
         let mut worst_ours = 0.0f64;
         let mut worst_naive = 0.0f64;
         // Sweep down to where `1 + x` stops being representable as distinct.
-        let xs: Vec<f64> = (0..320).map(|i| 0.5f64.powi(i as i32)).collect();
+        let xs: Vec<f64> = (0..320).map(|i| 0.5f64.powi(i)).collect();
         for &x in &xs {
             let want = x.ln_1p();
             if want == 0.0 {
